move contract clients
